[feat/#199] 컴포넌트화 작업
