+++ src/components/FormDetails/ActionButtons/ActionButton.module.scss
@@ -1,0 +1,26 @@
+@use 'src/styles/_media' as media;
+@use 'src/styles/_color' as color;
+@use 'src/styles/_font' as font;
+
+.button-container {
+  display: flex;
+  flex-direction: column;
+  row-gap: 16px;
+
+  @include media.tablet {
+    flex-direction: row-reverse;
+    gap: 0 8px;
+  }
+}
+
+.delete-tablet-button {
+  @include media.tablet {
+    width: 70px;
+  }
+}
+
+.no-text {
+  @include media.tablet {
+    display: none;
+  }
+}
