+++ src/components/FormDetails/ContactInfo/ContactInfo.tsx
@@ -1,18 +1,12 @@
 'use client'
 
-import MainButton from '@/components/Button/MainButton/MainButton'
-import AlertModal from '@/components/Modal/Alert/AlertModal'
-import {
-  useDeleteFormQuery,
-  useUsersMeQuery,
-} from '@/lib/queries/formDetailsQuery'
+import { useUsersMeQuery } from '@/lib/queries/formDetailsQuery'
 import { ContactInfoProps } from '@/lib/types/formTypes'
-import handleError from '@/lib/utils/errorHandler'
-import { formatPhoneNumber } from '@/lib/utils/formatDate'
 import { formatDate } from '@/lib/utils/dateFormatters'
-import { useRouter } from 'next/navigation'
+import { formatPhoneNumber } from '@/lib/utils/formatDate'
 import { useEffect, useState } from 'react'
 
+import ActionButtons from '../ActionButtons/ActionButtons'
 import styles from './ContactInfo.module.scss'
 
 export default function ContactInfo({
@@ -20,9 +14,7 @@ export default function ContactInfo({
 }: {
   formDetails: ContactInfoProps
 }) {
-  const router = useRouter()
   const { data: userRole } = useUsersMeQuery()
-  const { mutate: deleteForm } = useDeleteFormQuery()
   const recruitmentStartDate = formatDate.toKorean(
     formDetails?.recruitmentStartDate,
   )
@@ -30,10 +22,10 @@ export default function ContactInfo({
     formDetails?.recruitmentEndDate,
   )
   const isRecruitmentActive =
-    formDetails?.recruitmentEndDate &&
+    Boolean(formDetails?.recruitmentEndDate) &&
     new Date(formDetails.recruitmentEndDate) > new Date()
   const [statusMessage, setStatusMessage] = useState<string>('모집기간 계산 중')
-  const [isModalOpen, setIsModalOpen] = useState<boolean>(false)
+  const [isTablet, setIsTablet] = useState(false)
 
   useEffect(() => {
     if (recruitmentEndDate) {
@@ -50,49 +42,18 @@ export default function ContactInfo({
     }
   }, [recruitmentEndDate])
 
-  const handleApplyClick = () => {
-    router.push(`/form/${formDetails.id}/apply`)
+  const handleResize = () => {
+    setIsTablet(window.innerWidth < 1024)
   }
 
-  const handleShowApplicationHistory = () => {
-    router.push(`/form/${formDetails?.id}/application`)
-    // 지원자 -> 제출 내용 보기
-  }
-
-  const handleEditClick = () => {
-    router.push(`form/${formDetails.id}/edit`)
-  }
-
-  const handleDeleteClick = () => {
-    setIsModalOpen(true)
-  }
-
-  const closeModal = () => {
-    setIsModalOpen(false)
-  }
-
-  const handleConfirm = () => {
-    deleteForm(Number(formDetails?.id), {
-      onSuccess: () => {
-        router.push('/')
-        // 페이지네이션 목록으로 가기
-      },
-      onError: () => {
-        handleError(new Error('폼 삭제 실패'))
-      },
-    })
-  }
+  useEffect(() => {
+    handleResize()
+    window.addEventListener('resize', handleResize)
+    return () => window.removeEventListener('resize', handleResize)
+  }, [])
 
   return (
     <section className={styles['contact-info']}>
-      {isModalOpen && (
-        <AlertModal
-          AlertmodalType="delete"
-          isOpen={isModalOpen}
-          onRequestClose={closeModal}
-          onConfirm={handleConfirm}
-        />
-      )}
       <div className={styles['contact-info-container']}>
         <div
           className={`${styles['contact-info-auth']} ${styles['contact-info-line']}`}
@@ -126,51 +87,12 @@ export default function ContactInfo({
       </div>
 
       <div className={styles['button-container']}>
-        {userRole === 'APPLICANT' ? (
-          <>
-            <MainButton
-              buttonStyle="solid"
-              disabled={!isRecruitmentActive}
-              onClick={handleApplyClick}
-            >
-              <MainButton.Icon src="/icons/ic-writing.svg" altText="지원하기" />
-              <MainButton.Text>지원하기</MainButton.Text>
-            </MainButton>
-            <MainButton
-              buttonStyle="outline"
-              disabled={!isRecruitmentActive}
-              onClick={handleShowApplicationHistory}
-            >
-              <MainButton.Icon
-                src="/icons/ic-apply-list.svg"
-                altText="내 지원내역 보기"
-              />
-              <MainButton.Text>내 지원내역 보기</MainButton.Text>
-            </MainButton>
-          </>
-        ) : (
-          <>
-            <MainButton
-              buttonStyle="solid"
-              disabled={false}
-              onClick={handleEditClick}
-            >
-              <MainButton.Icon src="/icons/ic-edit2.svg" altText="수정하기" />
-              <MainButton.Text>수정하기</MainButton.Text>
-            </MainButton>
-            <MainButton
-              buttonStyle="outline"
-              disabled={false}
-              onClick={handleDeleteClick}
-              color="gray"
-            >
-              <MainButton.Icon
-                src="/icons/ic-trash-can.svg"
-                altText="삭제하기"
-              />
-              <MainButton.Text>삭제하기</MainButton.Text>
-            </MainButton>
-          </>
+        {!isTablet && (
+          <ActionButtons
+            userRole={userRole}
+            isRecruitmentActive={isRecruitmentActive}
+            formId={formDetails?.id}
+          />
         )}
       </div>
     </section>
