+++ src/components/Modal/Alert/AlertModal.module.scss
@@ -111,5 +111,8 @@
 .modal-alert-content-buttongroup {
   display: flex;
   flex-direction: column;
-  row-gap: 5px;
+}
+
+.main-button-border {
+  border: none;
 }
